refactor: remove redundant hooks and consolidate note logic into `Document` component
